Change `poll`'s timeout from `c_int` to `Option<&Timespec>`.
This harmonizes the timeout with the rest of rustix, which uses
`Timespec` for all time values. And, it eliminates the awkwardness
of using `-1` as a sentinel value.

On platforms with `ppoll`, the `Timespec` can be passed straight
to the OS. On platforms without `ppoll`, we have to do a fallible
conversion into `c_int`.